+++ web/apps/photos/src/components/PhotoViewer/FileInfo/index.tsx
@@ -2,8 +2,9 @@ import { EnteDrawer } from "@/base/components/EnteDrawer";
 import { Titlebar } from "@/base/components/Titlebar";
 import { nameAndExtension } from "@/base/file";
 import { FileType } from "@/media/file-type";
+import type { ParsedMetadata } from "@/media/types/file";
 import { UnidentifiedFaces } from "@/new/photos/components/PeopleList";
-import type { ParsedMetadata, RawExifTags } from "@/new/photos/services/exif";
+import type { RawExifTags } from "@/new/photos/services/exif";
 import { isMLEnabled } from "@/new/photos/services/ml";
 import { EnteFile } from "@/new/photos/types/file";
 import { formattedByteSize } from "@/new/photos/utils/units";
